+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Upgrade gjavah [#43](https://github.com/sbt/sbt-jni/pull/43)
 - Use cmake platform build tool [#40](https://github.com/sbt/sbt-jni/issues/40)
+- Rename macro project to core and simplify Scala 3 support [#52](https://github.com/sbt/sbt-jni/pull/52)
 
 ### Fixed
 - javah failed with ClassCastException [#38](https://github.com/sbt/sbt-jni/issues/38)